syllable-division-backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
django-cors-headers==3.14.0
python-dotenv==1.0.0
pytest==7.2.2
pytest-django==4.5.2
black==22.3.0
flake8==4.0.1